Renamed `gl_back_end.rs` to `back_end.rs`
Closes https://github.com/PistonDevelopers/opengl_graphics/issues/169
