+++ src/lib.rs
@@ -12,13 +12,13 @@ extern crate graphics;
 extern crate freetype;
 
 pub use shader_version::OpenGL;
-pub use gl_back_end::GlGraphics;
-pub use gl_back_end::GlGraphics as Gl;
+pub use back_end::GlGraphics;
+pub use back_end::GlGraphics as Gl;
 pub use texture::Texture;
 
 pub mod shader_utils;
 pub mod glyph_cache;
 pub mod error;
 
-mod gl_back_end;
+mod back_end;
 mod texture;
